Remove nodejs, bower
Because we're now committing the generated, cache busted files to the
repo, we don't need to compile them as part of travis.

We're committing them to the repo as we're deploying often enough that
we don't want to regenerate them all the time.
  - It takes time, and
  - It busts caches for no reason

Thus: We generate files as required, and commit them.