AD-HOC refactor (portfolio): Move work samples into a portfolio
Traditionally, a resume is used to express a snapshot of a potential job
candidates work history. However, given this particular vertical's
requirement to fluently understand HTML and CSS primitives and web
technology, such a resume does not provide the best opportunities to
understand a candidates skills and demonstrates of work.

Moving the work to a "portfolio" component allows a more complex
explanation of each component, with links to case studies and
demonstrations of the work in action.

This commit is the initial migration of the work from the resume to the
portfolio. Future work will expand on the portfolio, linking it to case
studies, blogs or other media and allowing the content to be filtered
based on the subject matter.