AD-HOC feat (JavaScript): Implement primitive JavaScript handling for errors
One of the tricker things to track as a software engineer is errors in
the client side of the application. Absent this visibility, the unaware
frontend engineer can mistakenly consider that their code is running
successfully, but users will end with a frustrating experience.

This commit implements a super quick an dirty exception tracking through
Google Analytics.