AD-HOC feat (Portfolio): Add skills to portfolio
In previous work some skills were added to each task to denote what each
portfolio item was supposed to denote. This commit renders those
prepared skills.

It additionally refactors the company, such that it's not implied that
it's filterable. Future work may revert this design, but it currently
does not make sense as the filters do not work.