AD-HOC fix (Resource): Add resources to sources list
Currently there is a problem in which the document form resume, added as
part of the previous commit, is not available in the production system.
This appears to be as it is not bundled in the build.

This commit adds it to the polymer sources configuration, such that it
should be included.