AD-HOC fix (Build): Make builds fail on single failure
Currently there is a problem in which an intermittent build failure (due
to a network dependency) fails, the build *does not* fail. This is a
design of the CI service used; Travis CI. Unfortunately this means there
can be a condition in which the deployment proceeds successfully, but on
an unsuccessful artifact.

This commit makes builds step dependent on each other, such that any
build failure will fail all steps, and the failure will propagate.

See
  - https://github.com/travis-ci/travis-ci/issues/1066
  - https://github.com/travis-ci/travis-build/pull/192