Generate the site with a Makefile
To resolve #37 (Add link to download copy of resume as PDF) there needs
to be a more complex build process.

There are many ways to encode this, but currently the application will
only be deployed for a short period and manual deployments are fine.

Given this, this commit creates a Makefile that automates the task of
generating the content. It uses the self documenting makefile style of
Enginyard and littlemanco/boilr-makefile