Add Google Analytics to the rendered page
Currently there is a problem in that it is difficult to determine how
successful any sort of campaign to spread the resume around will be.

This commit attempts to address that by adding Google Analytics to the
stack, which should measure the number of visitors to the website.

As the various links are posted through social media Google Analytics
should report in the various places people are interested in viewing.