Create the pattern of verifiable work
Currently there is a problem in which the resume has a little too much
information to be easily digestible. The problem is one of accuracy; The
goal is to provide the best possible information to a given user, but
provide them too much information and they will skip past the relevant
information.

This commit attempts to address that by providing multiple "paths"
outside the document that a given assessor can read further about, but
does not require that investigation on the first pass of the document.